refactor(utils): Move infer_family_from_object to utils module
Relocates `infer_family_from_object` function from validators to utils
module for better organization. Updates imports across affected modules
to reflect the new location.
