+++ netbox_acls/utils.py
@@ -8,7 +8,30 @@ from django.core.exceptions import ValidationError
 from django.db.backends.postgresql.psycopg_any import NumericRange
 from django.utils.translation import gettext_lazy as _
 
-__all__ = ("normalize_port_ranges",)
+from .choices import ACLFamilyChoices
+
+__all__ = (
+    "infer_family_from_object",
+    "normalize_port_ranges",
+)
+
+
+def infer_family_from_object(obj):
+    """
+    Infers the family type (IPv4 or IPv6) from a given object's attributes.
+    """
+    # Prefer a 'version' if present
+    version = (
+        getattr(obj, "family", None)
+        or getattr(getattr(obj, "prefix", None), "version", None)
+        or getattr(getattr(obj, "address", None), "version", None)
+        or getattr(getattr(obj, "start_address", None), "version", None)
+    )
+    if version == 4:
+        return ACLFamilyChoices.FAMILY_IPV4
+    if version == 6:
+        return ACLFamilyChoices.FAMILY_IPV6
+    return None
 
 
 def normalize_port_ranges(ranges: Iterable[NumericRange], field_name: str = "__all__") -> list[NumericRange]:
